changed dim_type typedef to int from long long
Using long long for indexing data type is causing drastic
performance drop for CUDA and OpenCL kernels. Hence, changing the
typedef to point to int.

+++ include/af/defines.h
@@ -97,7 +97,7 @@ typedef enum {
 
 typedef size_t af_array;
 
-typedef long long dim_type;
+typedef int dim_type;
 
 typedef enum {
 	AF_INTERP_NEAREST,
