+++ src/backend/opencl/kernel/bilateral.cl
@@ -24,8 +24,8 @@ void load2LocalMem(__local outType *  shrd,
         dim_type gx, dim_type gy,
         dim_type inStride1, dim_type inStride0)
 {
-    dim_type gx_  = clamp(gx, (long)0, dim0-1);
-    dim_type gy_  = clamp(gy, (long)0, dim1-1);
+    dim_type gx_  = clamp(gx, 0, dim0-1);
+    dim_type gy_  = clamp(gy, 0, dim1-1);
     shrd[ lIdx(lx, ly, shrdStride, 1) ] = (outType)in[ lIdx(gx_, gy_, inStride1, inStride0) ];
 }
